// Cloud5S_WEB/src/app/models/System-manage/system-parameter.model.ts
export interface SystemParameterModel {
  id: string;
  portName: string;
  portAddress: string;
  portLatitude: number;
  portLongitude: number;
  defaultItemExportCode: string;
  defaultItemImportCode: string;
  defaultIngredientItemCode: string;
  defaultProductItemCode: string;
  defaultIngredientStock: string;
  defaultProductStock: string;
  unitCode: string;
  defaultCompanyCode: string;
}